requirements.txt
google-adk
streamlit
pypdf
uuid